Standardize notifications page controller
